[#175266102] refactoring components
