+++ ts/features/bonus/bpd/screens/details/components/summary/base/ShadowBox.tsx
